Update ItemTimeInABottle.java
Now when there are multiple Time In A Bottle in player inventory, the one that has longest time will get its time increased, instead of letting others get resetted every 60 ticks. Note that this is different from original behavior, but it is still compatitable to switch between the two mods.